requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==25.1.0
requests==2.31.0
google-api-python-client==2.114.0
google-auth==2.27.0
